pylintrc: re-enable too-many-lines, part 4
To get under the 1000 line limit, move host_helpers
classes and methods to different files.
